Fix ExternalDataSupplier

+++ src/de/schafunschaf/bountiesexpanded/ExternalDataSupplier.java
@@ -49,7 +49,7 @@ public class ExternalDataSupplier {
             Settings.pirateBountyActive = settings.getBoolean("pirateBountyActive");
             Settings.disableVanillaBounties = settings.getBoolean("disableVanillaBounties");
             Settings.disableVayraBounties = settings.getBoolean("disableVayraBounties");
-            Settings.pirateBountySpawnChance = settings.getInt("pirateBountySpawnChance");
+            Settings.pirateBountySpawnChance = settings.getDouble("pirateBountySpawnChance");
             Settings.pirateBountyMinBounties = settings.getInt("pirateBountyMinBounties");
             Settings.pirateBountyMaxBounties = settings.getInt("pirateBountyMaxBounties");
             Settings.pirateBountyMinDuration = settings.getInt("pirateBountyMinDuration");
